backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-jose
passlib[bcrypt]
pydantic
bcrypt==4.0.1
python-jose
